User search with native query implemented
git-svn-id: https://svn.apache.org/repos/asf/incubator/syncope/trunk@1246734 13f79535-47bb-0310-9956-ffa450edef68

+++ client/src/main/java/org/syncope/client/search/PaginatedResult.java
@@ -2,9 +2,9 @@
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
  *  You may obtain a copy of the License at
- * 
+ *
  *       http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  *  Unless required by applicable law or agreed to in writing, software
  *  distributed under the License is distributed on an "AS IS" BASIS,
  *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -18,14 +18,11 @@ import java.util.List;
 import org.syncope.client.AbstractBaseBean;
 import org.syncope.client.to.UserTO;
 
-/**
- * Search condition to be applied when comparing attribute values.
- */
 public class PaginatedResult extends AbstractBaseBean {
 
     private List<UserTO> records;
 
-    private Long totalRecords;
+    private int totalRecords;
 
     private int recordsInPage;
 
@@ -65,11 +62,11 @@ public class PaginatedResult extends AbstractBaseBean {
         this.recordsInPage = recordsInPage;
     }
 
-    public Long getTotalRecords() {
+    public int getTotalRecords() {
         return totalRecords;
     }
 
-    public void setTotalRecords(Long totalRecords) {
+    public void setTotalRecords(int totalRecords) {
         this.totalRecords = totalRecords;
     }
 }
